Added: Flash when taking damage, Health and Speed UI
Also added TextMeshPro Asset Files

+++ Assets/Scripts/Player/Ship Stats Scriptable Objects/ShipStats.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New ShipStats", menuName = "ScriptableObjects/ShipStats", order = 1)]
 public class ShipStats : ScriptableObject
 {
-    public float hitPoints;
+    public int hitPoints;
 
     public float moveSpeed;
